prod/mysite/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.0
Django==5.1.4
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
docutils==0.21.2
drf-spectacular==0.28.0
flake8==7.2.0
flake8-docstrings==1.7.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pycodestyle==2.13.0
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.3.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sentry-sdk==2.30.0
snowballstemmer==3.0.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
